requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.31.0
pillow==9.5.0
python-jose[cryptography]==3.3.0
scipy==1.11.4 